+++ CHANGELOG.md
@@ -189,6 +189,8 @@ New functionality includes:
 
 ## NEW FUNCTIONALITY
 
+* `DockerPlatform`: add `target_platform` parameter to the docker platform config to enable multi-plaform builds.
+
 * `viash export json_schema`: Export a json schema derived from the class reflections and annotations already used by the `config_schema` (PR #446).
 
 * `viash export config_schema`: Output `default` values of member fields (PR #446).